编写wifi与车辆控制间的通信(转向间隔50ms,绝对值10时转向比较流畅)

+++ HARDWARE/servo/servo.h
@@ -3,5 +3,5 @@
 /* Includes ------------------------------------------------------------------*/
 #include "stm32f4xx_conf.h"
 extern void servo_init(void);
-extern void servo_ctrl(uint8_t value);
+extern void servo_ctrl(uint16_t value);
 #endif
